requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1

# LangChain + LangGraph
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langgraph>=0.2.15

# LLM providers (choose one)
langchain-groq>=0.1.0
langchain-google-genai>=0.1.0

# Embeddings + Vector store
sentence-transformers>=2.2.2
chromadb>=0.5.0

# Document loaders
pypdf>=4.2.0

# Web search tool
tavily-python>=0.3.8

# Testing
pytest>=8.2.0
httpx>=0.27.0
